# api.doc.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1.0.0",
    "title": " REST API"
  },
  "host": "localhost:8088",
  "basePath": "/",
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/user/changepwd": {
      "put": {
        "tags": [
          "user-controller"
        ],
        "summary": "Change the user password",
        "operationId": "changePasswordUsingPUT",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "in": "body",
            "name": "passwordChangeDTO",
            "description": "The new && old password",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PasswordChange"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully authenticated user",
            "schema": {
              "type": "string"
            }
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Some required fields are invalid"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "If the username or password is not recognized"
          }
        },
        "deprecated": false
      }
    },
    "/api/user/profile/update": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "updateProfile",
        "operationId": "updateProfileUsingPOST",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "in": "body",
            "name": "updateDTO",
            "description": "updateDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProfileUpdateDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseEntity"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/user/{username}/follow": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "follow",
        "operationId": "followUsingPOST",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "username",
            "in": "path",
            "description": "username",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseEntity"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/user/{username}/followed": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "isFollowed",
        "operationId": "isFollowedUsingGET",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "username",
            "in": "path",
            "description": "username",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseEntity"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/user/{username}/profile/base": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "Retrieve the basic profile information",
        "operationId": "getBasicInfoUsingGET",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "Username of the profile owner",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The profile data",
            "schema": {
              "$ref": "#/definitions/ProfileBasicInfo"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "The provided username does not exist"
          }
        },
        "deprecated": false
      }
    },
    "/api/user/{username}/profile/bookmarks": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "getBookmarks",
        "operationId": "getBookmarksUsingGET",
        "parameters": [
          {
            "name": "p",
            "in": "query",
            "description": "p",
            "required": false,
            "type": "integer",
            "default": 0,
            "format": "int32"
          },
          {
            "name": "username",
            "in": "path",
            "description": "username",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseEntity"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/user/{username}/profile/followers": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "Retrieve the user followers",
        "operationId": "getFollowersUsingGET",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Pagination",
            "required": false,
            "type": "integer",
            "default": 0,
            "format": "int32",
            "allowEmptyValue": false
          },
          {
            "name": "username",
            "in": "path",
            "description": "username",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The following users",
            "schema": {
              "$ref": "#/definitions/ProfileBasicInfo"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "The provided username does not exist"
          }
        },
        "deprecated": false
      }
    },
    "/api/user/{username}/profile/follows": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "Retrieve the users followed by an user",
        "operationId": "getFollowsUsingGET",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Pagination",
            "required": false,
            "type": "integer",
            "default": 0,
            "format": "int32",
            "allowEmptyValue": false
          },
          {
            "name": "username",
            "in": "path",
            "description": "username",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The followed users",
            "schema": {
              "$ref": "#/definitions/ProfileBasicInfo"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "The provided username does not exist"
          }
        },
        "deprecated": false
      }
    },
    "/api/user/{username}/profile/pro": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "getProInfo",
        "operationId": "getProInfoUsingGET",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "username",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseEntity"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/user/{username}/profile/social": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "Retrieve the social profile information",
        "operationId": "getSocialInfoUsingGET",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "Username of the profile owner",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The profile social data",
            "schema": {
              "$ref": "#/definitions/ProfileBasicInfo"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "The provided username does not exist"
          }
        },
        "deprecated": false
      }
    },
    "/api/user/{username}/unfollow": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "unfollow",
        "operationId": "unfollowUsingPOST",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "username",
            "in": "path",
            "description": "username",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseEntity"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
...
    "PasswordChange": {
      "type": "object",
      "properties": {
        "newPassword": {
          "type": "string"
        },
        "oldPassword": {
          "type": "string"
        }
      },
      "title": "PasswordChange",
      "description": "Model representing an PasswordChange"
    },
    "ProfileBasicInfo": {
      "type": "object",
      "properties": {
        "avatar_url": {
          "type": "string"
        },
        "domain": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "firstname": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "university": {
          "$ref": "#/definitions/UniversityInfo"
        }
      },
      "title": "ProfileBasicInfo",
      "description": "Model representing an user's profile basic information"
    },
    "ProfileUpdateDTO": {
      "type": "object",
      "properties": {
        "avatar_url": {
          "type": "string"
        },
        "bio": {
          "type": "string"
        },
        "domain": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "interests": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "research_groups": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "university": {
          "type": "string"
        }
      },
      "title": "ProfileUpdateDTO"
    },
    "ResponseEntity": {
      "type": "object",
      "properties": {
        "body": {
          "type": "object"
        },
        "statusCode": {
          "type": "string",
          "enum": [
            "100 CONTINUE",
            "101 SWITCHING_PROTOCOLS",
            "102 PROCESSING",
            "103 CHECKPOINT",
            "200 OK",
            "201 CREATED",
            "202 ACCEPTED",
            "203 NON_AUTHORITATIVE_INFORMATION",
            "204 NO_CONTENT",
            "205 RESET_CONTENT",
            "206 PARTIAL_CONTENT",
            "207 MULTI_STATUS",
            "208 ALREADY_REPORTED",
            "226 IM_USED",
            "300 MULTIPLE_CHOICES",
            "301 MOVED_PERMANENTLY",
            "302 FOUND",
            "302 MOVED_TEMPORARILY",
            "303 SEE_OTHER",
            "304 NOT_MODIFIED",
            "305 USE_PROXY",
            "307 TEMPORARY_REDIRECT",
            "308 PERMANENT_REDIRECT",
            "400 BAD_REQUEST",
            "401 UNAUTHORIZED",
            "402 PAYMENT_REQUIRED",
            "403 FORBIDDEN",
            "404 NOT_FOUND",
            "405 METHOD_NOT_ALLOWED",
            "406 NOT_ACCEPTABLE",
            "407 PROXY_AUTHENTICATION_REQUIRED",
            "408 REQUEST_TIMEOUT",
            "409 CONFLICT",
            "410 GONE",
            "411 LENGTH_REQUIRED",
            "412 PRECONDITION_FAILED",
            "413 PAYLOAD_TOO_LARGE",
            "413 REQUEST_ENTITY_TOO_LARGE",
            "414 URI_TOO_LONG",
            "414 REQUEST_URI_TOO_LONG",
            "415 UNSUPPORTED_MEDIA_TYPE",
            "416 REQUESTED_RANGE_NOT_SATISFIABLE",
            "417 EXPECTATION_FAILED",
            "418 I_AM_A_TEAPOT",
            "419 INSUFFICIENT_SPACE_ON_RESOURCE",
            "420 METHOD_FAILURE",
            "421 DESTINATION_LOCKED",
            "422 UNPROCESSABLE_ENTITY",
            "423 LOCKED",
            "424 FAILED_DEPENDENCY",
            "426 UPGRADE_REQUIRED",
            "428 PRECONDITION_REQUIRED",
            "429 TOO_MANY_REQUESTS",
            "431 REQUEST_HEADER_FIELDS_TOO_LARGE",
            "451 UNAVAILABLE_FOR_LEGAL_REASONS",
            "500 INTERNAL_SERVER_ERROR",
            "501 NOT_IMPLEMENTED",
            "502 BAD_GATEWAY",
            "503 SERVICE_UNAVAILABLE",
            "504 GATEWAY_TIMEOUT",
            "505 HTTP_VERSION_NOT_SUPPORTED",
            "506 VARIANT_ALSO_NEGOTIATES",
            "507 INSUFFICIENT_STORAGE",
            "508 LOOP_DETECTED",
            "509 BANDWIDTH_LIMIT_EXCEEDED",
            "510 NOT_EXTENDED",
            "511 NETWORK_AUTHENTICATION_REQUIRED"
          ]
        },
        "statusCodeValue": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "ResponseEntity"
    },
...
    "UniversityInfo": {
      "type": "object",
      "properties": {
        "abbreviation": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "website": {
          "type": "string"
        }
      },
      "title": "UniversityInfo",
      "description": "Model representing an information about an university"
    },
...
  }
}